brackets(#15): Updating ranking_bracket_system.js
Want to access information from indexeddb
