+++ team/front-end/source/brackets.js
@@ -1,6 +1,27 @@
 import { NavBarComponent } from './components/NavBarComponent/navbar.js';
+import { RankingBracketSystem } from './components/ranking_bracket_system/ranking_bracket_system.js';
+
 
 const app = document.getElementById('app');
 
 const navBar = new NavBarComponent();
 app.appendChild(navBar.render());
+
+
+document.addEventListener('DOMContentLoaded', () => {
+    const initialRestaurants = [
+        { name: 'Arigato', votes: 0 },
+        { name: 'Antonios', votes: 0 },
+        { name: 'Miss Saigon', votes: 0 },
+        { name: 'Chipotle', votes: 0 },
+        { name: 'McDonalds', votes: 0 },
+        { name: 'House of Teriyaki', votes: 0 },
+        { name: 'Bueno y Sano', votes: 0 },
+        { name: 'The Works', votes: 0 }
+    ];
+ 
+ 
+    const app = document.getElementById('app');
+    const rankingBracketSystem = new RankingBracketSystem(initialRestaurants);
+    app.appendChild(rankingBracketSystem.render());
+ });
